Subset of EFSA crop interception default values for groundwater modelling

Subset of EFSA crop interception default values for groundwater modelling

EFSA_GW_interception_2014

Format

A matrix containing interception values, currently only for some selected crops

Source

European Food Safety Authority (2014) EFSA Guidance Document for evaluating laboratory and field dissipation studies to obtain DegT50 values of active substances of plant protection products and transformation products of these active substances in soil. EFSA Journal 12(5):3662, 37 pp., doi:10.2903/j.efsa.2014.3662

Examples

EFSA_GW_interception_2014
#>                            BBCH
#> Crop                        0x   1x   2x   3x   4x   5x   6x   7x   8x   9x
#>   Beans (field + vegetable)  0 0.25 0.40 0.40 0.70 0.70 0.70 0.70 0.70 0.80
#>   Peas                       0 0.35 0.55 0.55 0.85 0.85 0.85 0.85 0.85 0.85
#>   Summer oilseed rape        0 0.40 0.80 0.80 0.80 0.80 0.80 0.80 0.80 0.90
#>   Winter oilseed rape        0 0.40 0.80 0.80 0.80 0.80 0.80 0.80 0.80 0.90
#>   Tomatoes                   0 0.50 0.70 0.70 0.80 0.80 0.80 0.80 0.80 0.50
#>   Spring cereals             0 0.00 0.20 0.80 0.90 0.90 0.90 0.80 0.80 0.80
#>   Winter cereals             0 0.00 0.20 0.80 0.90 0.90 0.90 0.80 0.80 0.80