FOCUS PELMO crop acronyms used for naming sceneario files

A named character vector with the crop acronyms used in FOCUS PELMO 5.5.3 for naming the .sze files located in the FOCUS directory. The crop acronyms in the file names are sometimes capitalized, sometimes not. The scenario files used for Beans (field) and Beans (vegetable) are the same.

FOCUS_PELMO_crop_sze_names

Format

An object of class character of length 25.

Examples

print(FOCUS_PELMO_crop_sze_names)

#>      app      gra      pot      sug      win      fbe      vbe      bus 
#> "apples"  "grass" "potato" "sbeets" "wcerea"  "beans"  "beans"  "bushb" 
#>      cab      car      cit      cot      lin      mai      soi      woi 
#> "cabbag" "carrot" "citrus" "cotton"  "linse"  "maize" "rapesu" "rapewi" 
#>      oni      ape      soy      spr      str      sun      tob      tom 
#> "onions"   "peas"   "soyb" "scerea" "strawb" "sunflo" "tobacc" "tomato" 
#>      vin 
#>  "vines"