FOCUS PELMO crop names

A named character vector with the crop names used in the PELMO 5.5.3 GUI. For the names, three letter codes were constructed by generally taking the first three letters in lower case. Only when there is an expression in parentheses, the first letter of this expression becomes the first letter in the three letter code, i.e. 'Peas (animals)' has the code ape.

FOCUS_PELMO_crops

Format

An object of class character of length 25.

Examples

print(FOCUS_PELMO_crops)

#>                      app                      gra                      pot 
#>                 "Apples"      "Grass and alfalfa"               "Potatoes" 
#>                      sug                      win                      fbe 
#>            "Sugar beets"         "Winter cereals"          "Beans (field)" 
#>                      vbe                      bus                      cab 
#>     "Beans (vegetables)"            "Bushberries"                "Cabbage" 
#>                      car                      cit                      cot 
#>                "Carrots"                 "Citrus"                 "Cotton" 
#>                      lin                      mai                      soi 
#>                "Linseed"                  "Maize" "Oil seed rape (summer)" 
#>                      woi                      oni                      ape 
#> "Oil seed rape (winter)"                 "Onions"         "Peas (animals)" 
#>                      soy                      spr                      str 
#>               "Soybeans"         "Spring cereals"           "Strawberries" 
#>                      sun                      tob                      tom 
#>              "Sunflower"                "Tobacco"               "Tomatoes" 
#>                      vin 
#>                  "Vines"