Set non-detects in residue series without replicates

Sets non-detects directly before or directly after detects to NA. Values between lod and loq are set to their mean value if an loq is specified. If 'time_zero' is set to TRUE, the residue series is assumed to start with time zero, and non-detects at time zero are set to 'time_zero_nd_value'. For the set_nd_focus variant, this is zero, otherwise this argument has NA as default value. If stopping after the first non-detection is requested, as in in the FOCUS variant of the function, an loq has to be specified in order to decide if any later detections are above the loq.

set_nd(r, lod, loq = NA, time_zero = TRUE, time_zero_nd_value = NA,
  stop_after_first_nondetect = FALSE)

set_nd_focus(r, lod, loq = NA, time_zero = TRUE)

Arguments

r
A character vector of sequential residues without replicates, with non-detects specified as 'nd' and unquantified values above the limit of detection specified as 'nq', otherwise coercible to numeric

lod
Limit of detection (numeric)

loq
Limit of quantification(numeric). Must be specified if the FOCUS rule to stop after the first non-detection is to be applied

time_zero
Is the first value in the series a time zero value?

time_zero_nd_value
Which value should we use for non-detects at time zero?

stop_after_first_nondetect
Should we really stop after the first non-detection?

Functions

set_nd: Set non-detects in residues series

set_nd_focus: Set non-detects in residues series according to FOCUS rules

References

FOCUS (2014) Generic Guidance for Estimating Persistence and Degradation Kinetics from Environmental Fate Studies on Pesticides in EU Registration, Version 1.1, 18 December 2014, p. 251

Examples

# FOCUS (2014) p. 75/76 and 131/132
parent_1 <- c(.12, .09, .05, .03, "nd", "nd", "nd", "nd", "nd", "nd")
set_nd(parent_1, 0.02)

#>  [1] 0.12 0.09 0.05 0.03 0.01   NA   NA   NA   NA   NA

parent_2 <- c(.12, .09, .05, .03, "nd", "nd", .03, "nd", "nd", "nd")
set_nd(parent_2, 0.02)

#>  [1] 0.12 0.09 0.05 0.03 0.01 0.01 0.03 0.01   NA   NA

set_nd_focus(parent_2, 0.02, loq = 0.05)

#>  [1] 0.12 0.09 0.05 0.03 0.01   NA   NA   NA   NA   NA

parent_3 <- c(.12, .09, .05, .03, "nd", "nd", .06, "nd", "nd", "nd")
set_nd(parent_3, 0.02)

#>  [1] 0.12 0.09 0.05 0.03 0.01 0.01 0.06 0.01   NA   NA

set_nd_focus(parent_3, 0.02, loq = 0.05)

#>  [1] 0.12 0.09 0.05 0.03 0.01 0.01 0.06 0.01   NA   NA

metabolite <- c("nd", "nd", "nd", 0.03, 0.06, 0.10, 0.11, 0.10, 0.09, 0.05, 0.03, "nd", "nd")
set_nd(metabolite, 0.02)

#>  [1]   NA   NA 0.01 0.03 0.06 0.10 0.11 0.10 0.09 0.05 0.03 0.01   NA