Properties of the predefined scenarios from the EFSA guidance from 2017

Properties of the predefined scenarios used at Tier 1, Tier 2A and Tier 3A for the concentration in soil as given in the EFSA guidance (2017, p. 14/15). Also, the scenario and model adjustment factors from p. 16 and p. 18 are included.

Format

A data frame with one row for each scenario. Row names are the scenario codes, e.g. CTN for the Northern scenario for the total concentration in soil. Columns are mostly self-explanatory. rho is the dry bulk density of the top soil.

Source

EFSA (European Food Safety Authority) (2017) EFSA guidance document for predicting environmental concentrations of active substances of plant protection products and transformation products of these active substances in soil. EFSA Journal 15(10) 4982 doi:10.2903/j.efsa.2017.4982

Examples

soil_scenario_data_EFSA_2017

#>        Zone Country T_arit T_arr Texture  f_om theta_fc   rho f_sce f_mod
#> CTN   North Estonia    5.7   7.6  Coarse 0.220    0.244 0.707   1.4     3
#> CTC Central  Poland    7.4   9.3  Coarse 0.122    0.244 0.934   1.4     3
#> CTS   South  France   10.2  11.7  Medium 0.070    0.349 1.117   1.4     3
#> CLN   North Denmark    8.0   9.2  Medium 0.025    0.349 1.371   1.6     4
#> CLC Central Austria    9.3  11.3  Medium 0.018    0.349 1.432   1.6     4
#> CLS   South   Spain   15.4  16.7  Medium 0.010    0.349 1.521   1.6     4
#>     FOCUS_zone prec
#> CTN    Hamburg  639
#> CTC    Hamburg  617
#> CTS    Hamburg  667
#> CLN    Hamburg  602
#> CLC Châteaudun  589
#> CLS    Sevilla  526