Subset of EFSA crop washoff default values

Subset of EFSA crop washoff default values

Format

A matrix containing wash-off factors, currently only for some selected crops

Source

European Food Safety Authority (2017) EFSA guidance document for predicting environmental concentrations of active substances of plant protection products and transformation products of these active substances in soil. EFSA Journal 15(10) 4982 doi:10.2903/j.efsa.2017.4982

Examples

if (FALSE) {
  # This is the code that was used to define the data
  bbch <- paste0(0:9, "x")
  crops <- c(
    "Beans (field + vegetable)",
    "Peas",
    "Summer oilseed rape", "Winter oilseed rape",
    "Tomatoes",
    "Spring cereals", "Winter cereals")
  EFSA_washoff_2017 <- matrix(NA, length(crops), length(bbch),
    dimnames = list(Crop = crops, BBCH = bbch))
  EFSA_washoff_2017["Beans (field + vegetable)", ] <-
    c(NA, 0.6, rep(0.75, 2), rep(0.8, 5), 0.35)
  EFSA_washoff_2017["Peas", ] <-
    c(NA, 0.4, rep(0.6, 2), rep(0.65, 5), 0.35)
  EFSA_washoff_2017["Summer oilseed rape", ] <-
    c(NA, 0.4, rep(0.5, 2), rep(0.6, 5), 0.5)
  EFSA_washoff_2017["Winter oilseed rape", ] <-
    c(NA, 0.1, rep(0.4, 2), rep(0.55, 5), 0.3)
  EFSA_washoff_2017["Tomatoes", ] <-
    c(NA, 0.55, rep(0.75, 2), rep(0.7, 5), 0.35)
  EFSA_washoff_2017["Spring cereals", ] <-
    c(NA, 0.4, 0.5, 0.5, rep(0.65, 3), rep(0.65, 2), 0.55)
  EFSA_washoff_2017["Winter cereals", ] <-
    c(NA, 0.1, 0.4, 0.6, rep(0.55, 3), rep(0.6, 2), 0.4)
  save(EFSA_washoff_2017,
    file = "../data/EFSA_washoff_2017.RData")
}
EFSA_washoff_2017
#>                            BBCH
#> Crop                        0x   1x   2x   3x   4x   5x   6x   7x   8x   9x
#>   Beans (field + vegetable) NA 0.60 0.75 0.75 0.80 0.80 0.80 0.80 0.80 0.35
#>   Peas                      NA 0.40 0.60 0.60 0.65 0.65 0.65 0.65 0.65 0.35
#>   Summer oilseed rape       NA 0.40 0.50 0.50 0.60 0.60 0.60 0.60 0.60 0.50
#>   Winter oilseed rape       NA 0.10 0.40 0.40 0.55 0.55 0.55 0.55 0.55 0.30
#>   Tomatoes                  NA 0.55 0.75 0.75 0.70 0.70 0.70 0.70 0.70 0.35
#>   Spring cereals            NA 0.40 0.50 0.50 0.65 0.65 0.65 0.65 0.65 0.55
#>   Winter cereals            NA 0.10 0.40 0.60 0.55 0.55 0.55 0.60 0.60 0.40