<<<<<<< HEAD

Step 1/2 scenario data as distributed with the FOCUS Step 1/2 calculator
=======

Step 1/2 scenario data as distributed with the FOCUS Step 1/2 calculator

The data were extracted from the scenario.txt file using the R code shown below. The text file is not included in the package as its licence is not clear.

<<<<<<< HEAD

Format

A list containing the scenario names in a character vector called 'names', the drift percentiles in a matrix called 'drift', interception percentages in a matrix called 'interception' and the runoff/drainage percentages for Step 2 calculations in a matrix called 'rd'.

Examples

# \dontrun{
  # This is the code that was used to extract the data
  scenario_path <- "inst/extdata/FOCUS_Step_12_scenarios.txt"
  scenarios <- readLines(scenario_path)[9:38]
#> Warning: cannot open file 'inst/extdata/FOCUS_Step_12_scenarios.txt': No such file or directory
#> Error in file(con, "r"): cannot open the connection
  FOCUS_Step_12_scenarios <- list()
  sce <- read.table(text = scenarios, sep = "\t", header = TRUE, check.names = FALSE,
    stringsAsFactors = FALSE)
#> Error in eval(expr, envir, enclos): object 'scenarios' not found
  FOCUS_Step_12_scenarios$names = sce$Crop
#> Error in eval(expr, envir, enclos): object 'sce' not found
  rownames(sce) <- sce$Crop
#> Error in eval(expr, envir, enclos): object 'sce' not found
  FOCUS_Step_12_scenarios$drift = sce[, 3:11]
#> Error in eval(expr, envir, enclos): object 'sce' not found
  FOCUS_Step_12_scenarios$interception = sce[, 12:15]
#> Error in eval(expr, envir, enclos): object 'sce' not found
  sce_2 <- readLines(scenario_path)[41:46]
#> Warning: cannot open file 'inst/extdata/FOCUS_Step_12_scenarios.txt': No such file or directory
#> Error in file(con, "r"): cannot open the connection
  rd <- read.table(text = sce_2, sep = "\t")[1:2]
#> Error in eval(expr, envir, enclos): object 'sce_2' not found
  rd_mat <- matrix(rd$V2, nrow = 3, byrow = FALSE)
#> Error in eval(expr, envir, enclos): object 'rd' not found
  dimnames(rd_mat) = list(Time = c("Oct-Feb", "Mar-May", "Jun-Sep"),
                          Region = c("North", "South"))
#> Error: object 'rd_mat' not found
  FOCUS_Step_12_scenarios$rd = rd_mat
#> Error in eval(expr, envir, enclos): object 'rd_mat' not found
  save(FOCUS_Step_12_scenarios, file = "data/FOCUS_Step_12_scenarios.RData")
#> Warning: cannot open compressed file 'data/FOCUS_Step_12_scenarios.RData', probable reason 'No such file or directory'
#> Error in gzfile(file, "wb"): cannot open the connection
# }

# And this is the resulting data
FOCUS_Step_12_scenarios
#> list()

=======

Format

A list containing the scenario names in a character vector called 'names', the drift percentiles in a matrix called 'drift', interception percentages in a matrix called 'interception' and the runoff/drainage percentages for Step 2 calculations in a matrix called 'rd'.

Examples

if (FALSE) {
  # This is the code that was used to extract the data
  scenario_path <- "inst/extdata/FOCUS_Step_12_scenarios.txt"
  scenarios <- readLines(scenario_path)[9:38]
  FOCUS_Step_12_scenarios <- list()
  sce <- read.table(text = scenarios, sep = "\t", header = TRUE, check.names = FALSE,
    stringsAsFactors = FALSE)
  FOCUS_Step_12_scenarios$names = sce$Crop
  rownames(sce) <- sce$Crop
  FOCUS_Step_12_scenarios$drift = sce[, 3:11]
  FOCUS_Step_12_scenarios$interception = sce[, 12:15]
  sce_2 <- readLines(scenario_path)[41:46]
  rd <- read.table(text = sce_2, sep = "\t")[1:2]
  rd_mat <- matrix(rd$V2, nrow = 3, byrow = FALSE)
  dimnames(rd_mat) = list(Time = c("Oct-Feb", "Mar-May", "Jun-Sep"),
                          Region = c("North", "South"))
  FOCUS_Step_12_scenarios$rd = rd_mat
  save(FOCUS_Step_12_scenarios, file = "data/FOCUS_Step_12_scenarios.RData")
}

# And this is the resulting data
FOCUS_Step_12_scenarios
#> $names
#>  [1] "cereals, spring"                  "cereals, winter"                 
#>  [3] "citrus"                           "cotton"                          
#>  [5] "field beans"                      "grass / alfalfa"                 
#>  [7] "hops"                             "legumes"                         
#>  [9] "maize"                            "oil seed rape, spring"           
#> [11] "oil seed rape, winter"            "olives"                          
#> [13] "pome / stone fruit, early applns" "pome / stone fruit, late applns" 
#> [15] "potatoes"                         "soybeans"                        
#> [17] "sugar beets"                      "sunflowers"                      
#> [19] "tobacco"                          "vegetables, bulb"                
#> [21] "vegetables, fruiting"             "vegetables, leafy"               
#> [23] "vegetables, root"                 "vines, early applns"             
#> [25] "vines, late applns"               "appln, aerial"                   
#> [27] "appln, hand (crop < 50 cm)"       "appln, hand (crop > 50 cm)"      
#> [29] "no drift (incorp or seed trtmt)" 
#> 
#> $drift
#>                                       1      2      3      4      5      6
#> cereals, spring                   2.759  2.438  2.024  1.862  1.794  1.631
#> cereals, winter                   2.759  2.438  2.024  1.862  1.794  1.631
#> citrus                           15.725 12.129 11.011 10.124  9.743  9.204
#> cotton                            2.759  2.438  2.024  1.862  1.794  1.631
#> field beans                       2.759  2.438  2.024  1.862  1.794  1.631
#> grass / alfalfa                   2.759  2.438  2.024  1.862  1.794  1.631
#> hops                             19.326 17.723 15.928 15.378 15.114 14.902
#> legumes                           2.759  2.438  2.024  1.862  1.794  1.631
#> maize                             2.759  2.438  2.024  1.862  1.794  1.631
#> oil seed rape, spring             2.759  2.438  2.024  1.862  1.794  1.631
#> oil seed rape, winter             2.759  2.438  2.024  1.862  1.794  1.631
#> olives                           15.725 12.129 11.011 10.124  9.743  9.204
#> pome / stone fruit, early applns 29.197 25.531 23.960 23.603 23.116 22.760
#> pome / stone fruit, late applns  15.725 12.129 11.011 10.124  9.743  9.204
#> potatoes                          2.759  2.438  2.024  1.862  1.794  1.631
#> soybeans                          2.759  2.438  2.024  1.862  1.794  1.631
#> sugar beets                       2.759  2.438  2.024  1.862  1.794  1.631
#> sunflowers                        2.759  2.438  2.024  1.862  1.794  1.631
#> tobacco                           2.759  2.438  2.024  1.862  1.794  1.631
#> vegetables, bulb                  2.759  2.438  2.024  1.862  1.794  1.631
#> vegetables, fruiting              2.759  2.438  2.024  1.862  1.794  1.631
#> vegetables, leafy                 2.759  2.438  2.024  1.862  1.794  1.631
#> vegetables, root                  2.759  2.438  2.024  1.862  1.794  1.631
#> vines, early applns               2.699  2.496  2.546  2.499  2.398  2.336
#> vines, late applns                8.028  7.119  6.898  6.631  6.636  6.431
#> appln, aerial                    33.200 33.200 33.200 33.200 33.200 33.200
#> appln, hand (crop < 50 cm)        2.759  2.438  2.024  1.862  1.794  1.631
#> appln, hand (crop > 50 cm)        8.028  7.119  6.898  6.631  6.636  6.431
#> no drift (incorp or seed trtmt)   0.000  0.000  0.000  0.000  0.000  0.000
#>                                       7      8     >8
#> cereals, spring                   1.578  1.512  1.512
#> cereals, winter                   1.578  1.512  1.512
#> citrus                            9.102  8.656  8.656
#> cotton                            1.578  1.512  1.512
#> field beans                       1.578  1.512  1.512
#> grass / alfalfa                   1.578  1.512  1.512
#> hops                             14.628 13.520 13.520
#> legumes                           1.578  1.512  1.512
#> maize                             1.578  1.512  1.512
#> oil seed rape, spring             1.578  1.512  1.512
#> oil seed rape, winter             1.578  1.512  1.512
#> olives                            9.102  8.656  8.656
#> pome / stone fruit, early applns 22.690 22.241 22.241
#> pome / stone fruit, late applns   9.102  8.656  8.656
#> potatoes                          1.578  1.512  1.512
#> soybeans                          1.578  1.512  1.512
#> sugar beets                       1.578  1.512  1.512
#> sunflowers                        1.578  1.512  1.512
#> tobacco                           1.578  1.512  1.512
#> vegetables, bulb                  1.578  1.512  1.512
#> vegetables, fruiting              1.578  1.512  1.512
#> vegetables, leafy                 1.578  1.512  1.512
#> vegetables, root                  1.578  1.512  1.512
#> vines, early applns               2.283  2.265  2.265
#> vines, late applns                6.227  6.173  6.173
#> appln, aerial                    33.200 33.200 33.200
#> appln, hand (crop < 50 cm)        1.578  1.512  1.512
#> appln, hand (crop > 50 cm)        6.227  6.173  6.173
#> no drift (incorp or seed trtmt)   0.000  0.000  0.000
#> 
#> $interception
#>                                  no interception minimal crop cover
#> cereals, spring                                0               0.00
#> cereals, winter                                0               0.00
#> citrus                                         0               0.80
#> cotton                                         0               0.30
#> field beans                                    0               0.25
#> grass / alfalfa                                0               0.40
#> hops                                           0               0.20
#> legumes                                        0               0.25
#> maize                                          0               0.25
#> oil seed rape, spring                          0               0.40
#> oil seed rape, winter                          0               0.40
#> olives                                         0               0.70
#> pome / stone fruit, early applns               0               0.20
#> pome / stone fruit, late applns                0               0.20
#> potatoes                                       0               0.15
#> soybeans                                       0               0.20
#> sugar beets                                    0               0.20
#> sunflowers                                     0               0.20
#> tobacco                                        0               0.20
#> vegetables, bulb                               0               0.10
#> vegetables, fruiting                           0               0.25
#> vegetables, leafy                              0               0.25
#> vegetables, root                               0               0.25
#> vines, early applns                            0               0.40
#> vines, late applns                             0               0.40
#> appln, aerial                                  0               0.20
#> appln, hand (crop < 50 cm)                     0               0.20
#> appln, hand (crop > 50 cm)                     0               0.20
#> no drift (incorp or seed trtmt)                0               0.00
#>                                  average crop cover full canopy
#> cereals, spring                                0.20        0.70
#> cereals, winter                                0.20        0.70
#> citrus                                         0.80        0.80
#> cotton                                         0.60        0.75
#> field beans                                    0.40        0.70
#> grass / alfalfa                                0.60        0.75
#> hops                                           0.50        0.70
#> legumes                                        0.50        0.70
#> maize                                          0.50        0.75
#> oil seed rape, spring                          0.70        0.75
#> oil seed rape, winter                          0.70        0.75
#> olives                                         0.70        0.70
#> pome / stone fruit, early applns               0.40        0.65
#> pome / stone fruit, late applns                0.40        0.65
#> potatoes                                       0.50        0.70
#> soybeans                                       0.50        0.75
#> sugar beets                                    0.70        0.75
#> sunflowers                                     0.50        0.75
#> tobacco                                        0.70        0.75
#> vegetables, bulb                               0.25        0.40
#> vegetables, fruiting                           0.50        0.70
#> vegetables, leafy                              0.40        0.70
#> vegetables, root                               0.50        0.70
#> vines, early applns                            0.50        0.60
#> vines, late applns                             0.50        0.60
#> appln, aerial                                  0.50        0.70
#> appln, hand (crop < 50 cm)                     0.50        0.70
#> appln, hand (crop > 50 cm)                     0.50        0.70
#> no drift (incorp or seed trtmt)                0.00        0.00
#> 
#> $rd
#>          Region
#> Time      North South
#>   Oct-Feb     5     4
#>   Mar-May     2     4
#>   Jun-Sep     2     3
#> 

>>>>>>> refs/remotes/origin/main <<<<<<< HEAD ======= >>>>>>> refs/remotes/origin/main