Calculate initial predicted environmental concentrations in surface water due to drainage using the UK method

This implements the method specified in the UK data requirements handbook and was checked against the spreadsheet published on the CRC website. Degradation before the start of the drainage period is taken into account if latest_application is specified and the degradation parameters are given either as a soil_DT50 or a model.

Usage

PEC_sw_drainage_UK(
  rate,
  interception = 0,
  Koc,
  latest_application = NULL,
  soil_DT50 = NULL,
  model = NULL,
  model_parms = NULL
)

Arguments

rate

Application rate in g/ha

interception

The fraction of the application rate that does not reach the soil

Koc

The sorption coefficient normalised to organic carbon in L/kg

latest_application

Latest application date, formatted as e.g. "01 July"

soil_DT50

Soil degradation half-life, if SFO kinetics are to be used

model

The soil degradation model to be used. Either one of "FOMC", "DFOP", "HS", or "IORE", or an mkinmod object

model_parms

A named numeric vector containing the model parameters

Value

The predicted concentration in surface water in µg/L

References

HSE's Chemicals Regulation Division (CRD) Active substance PECsw calculations (for UK specific authorisation requests) https://www.hse.gov.uk/pesticides/topics/pesticide-approvals/pesticides-registration/data-requirements-handbook/fate/active-substance-uk.htm accessed 2019-09-27

Drainage PECs Version 1.0 (2015) Spreadsheet published at https://www.hse.gov.uk/pesticides/topics/pesticide-approvals/pesticides-registration/data-requirements-handbook/fate/pec-tools-2015/PEC%20sw-sed%20(drainage).xlsx accessed 2019-09-27

Author

Johannes Ranke

Examples

PEC_sw_drainage_UK(150, Koc = 100)
#> [1] 8.076923
PEC_sw_drainage_UK(60, interception = 0.5, Koc = 550,
  latest_application = "01 July", soil_DT50 = 200)
#> [1] 0.8388303