Calculate initial predicted environmental concentrations in surface water due to drainage using the UK method

This implements the method specified in the UK data requirements handbook and was checked against the spreadsheet published on the CRC website

PEC_sw_drainage_UK(rate, interception = 0, Koc,
  latest_application = NULL, soil_DT50 = NULL, model = NULL,
  model_parms = NULL)

Arguments

rate
Application rate in g/ha

interception
The fraction of the application rate that does not reach the soil

Koc
The sorption coefficient normalised to organic carbon in L/kg

latest_application
Latest application date, formatted as e.g. "01 July"

soil_DT50
Soil degradation half-life, if SFO kinetics are to be used

model
The soil degradation model to be used. Either one of "FOMC", "DFOP", "HS", or "IORE", or an mkinmod object

model_parms
A named numeric vector containing the model parameters

Value

The predicted concentration in surface water in µg/L

Examples

PEC_sw_drainage_UK(150, Koc = 100)

#> [1] 8.076923