Calculate predicted environmental concentrations in surface water due to drift

This is a basic, vectorised form of a simple calculation of a contaminant concentration in surface water based on complete, instantaneous mixing with input via spray drift.

Usage

PEC_sw_drift(
  rate,
  applications = 1,
  water_depth = 30,
  drift_percentages = NULL,
  drift_data = c("JKI", "RF"),
  crop_group_JKI = c("Ackerbau", "Obstbau frueh", "Obstbau spaet", "Weinbau frueh",
    "Weinbau spaet", "Hopfenbau", "Flaechenkulturen > 900 l/ha", "Gleisanlagen"),
  crop_group_focus = c("arable", "hops", "vines, late", "vines, early", "fruit, late",
    "fruit, early", "aerial"),
  distances = c(1, 5, 10, 20),
  rate_units = "g/ha",
  PEC_units = "µg/L"
)

Arguments

rate

Application rate in units specified below

applications

Number of applications for selection of drift percentile

water_depth

Depth of the water body in cm

drift_percentages

Percentage drift values for which to calculate PECsw. 'drift_data' and 'distances' if not NULL.

drift_data

Source of drift percentage data. If 'JKI', the drift_data_JKI included in the package is used. If 'RF', the Rautmann formula is used, if implemented for the crop type and number of applications

crop_group_JKI

When using the 'JKI' drift data, one of the German names as used in drift_parameters_focus.

crop_group_focus

One of the crop groups as used in drift_parameters_focus

distances

The distances in m for which to get PEC values

rate_units

Defaults to g/ha

PEC_units

Requested units for the calculated PEC. Only µg/L currently supported

Value

The predicted concentration in surface water

See also

Author

Johannes Ranke

Examples

PEC_sw_drift(100)
#>        1 m        5 m       10 m       20 m 
#> 0.92333333 0.19000000 0.09666667 0.05000000 
# Alternatively, we can use the formula for a single application to "Ackerbau" from the paper
PEC_sw_drift(100, drift_data = "RF")
#>        1 m        5 m       10 m       20 m 
#> 0.91976667 0.19064473 0.09680051 0.04915079 

# This makes it possible to also use different distances
PEC_sw_drift(100, distances = c(1, 3, 5, 6, 10, 20, 50, 100), drift_data = "RF")
#>        1 m        3 m        5 m        6 m       10 m       20 m       50 m 
#> 0.91976667 0.31415827 0.19064473 0.15951494 0.09680051 0.04915079 0.02006434 
#>      100 m 
#> 0.01018774 

# or consider aerial application
PEC_sw_drift(100, distances = c(1, 3, 5, 6, 10, 20, 50, 100), drift_data = "RF", 
  crop_group_focus = "aerial")
#>        1 m        3 m        5 m        6 m       10 m       20 m       50 m 
#> 16.8233333 11.0585820  9.0986174  8.4866460  6.9825178  4.7004640  1.8820816 
#>      100 m 
#>  0.9417586 

# Using custom drift percentages is also supported
PEC_sw_drift(100, drift_percentages = c(2.77, 0.95, 0.57, 0.48, 0.29, 0.15, 0.06, 0.03))
#>     2.77 %     0.95 %     0.57 %     0.48 %     0.29 %     0.15 %     0.06 % 
#> 0.92333333 0.31666667 0.19000000 0.16000000 0.09666667 0.05000000 0.02000000 
#>     0.03 % 
#> 0.01000000