<<<<<<< HEAD

Calculate PEC surface water due to runoff and erosion as in Exposit 3
=======

Calculate PEC surface water due to runoff and erosion as in Exposit 3

This is a reimplementation of the calculation described in the Exposit 3.02 spreadsheet file, in the worksheet "Konzept Runoff".

<<<<<<< HEAD
=======

Usage
>>>>>>> refs/remotes/origin/main

PEC_sw_exposit_runoff(
  rate,
  interception = 0,
  Koc,
  DT50 = Inf,
  t_runoff = 3,
  exposit_reduction_version = c("3.02", "3.01a", "3.01a2", "2.0"),
  V_ditch = 30,
  V_event = 100,
  dilution = 2
)

<<<<<<< HEAD

Source

Excel 3.02 spreadsheet available from https://www.bvl.bund.de/SharedDocs/Downloads/04_Pflanzenschutzmittel/zul_umwelt_exposit.html

Arguments
=======

Source

Excel 3.02 spreadsheet available from https://www.bvl.bund.de/DE/04_Pflanzenschutzmittel/03_Antragsteller/04_Zulassungsverfahren/07_Naturhaushalt/psm_naturhaush_node.html#doc1400590bodyText3

Arguments
>>>>>>> refs/remotes/origin/main

rate

The application rate in g/ha

interception

The fraction intercepted by the crop

Koc

The sorption coefficient to soil organic carbon

DT50

The soil half-life in days

t_runoff

The time between application and the runoff event, where degradation occurs, in days

exposit_reduction_version

The version of the reduction factors to be used. "3.02" is the current version used in Germany, "3.01a" is the version with additional percentages for 3 m and 6 m buffer zones used in Switzerland. "3.01a2" is a version introduced for consistency with previous calculations performed for a 3 m buffer zone in Switzerland, with the same reduction being applied to the dissolved and the bound fraction.

V_ditch

The volume of the ditch is assumed to be 1 m * 100 m * 30 cm = 30 m3

V_event

The unreduced runoff volume, equivalent to 10 mm precipitation on 1 ha

dilution

The dilution factor

<<<<<<< HEAD

Value

A list containing the following components

perc_runoff

The runoff percentages for dissolved and bound substance

runoff

A matrix containing dissolved and bound input for the different distances

PEC_sw_runoff

A matrix containing PEC values for dissolved and bound substance for the different distances. If the rate was given in g/ha, the PECsw are in microg/L.

See also

Examples
=======

Value

A list containing the following components

perc_runoff

The runoff percentages for dissolved and bound substance

runoff

A matrix containing dissolved and bound input for the different distances

PEC_sw_runoff

A matrix containing PEC values for dissolved and bound substance for the different distances. If the rate was given in g/ha, the PECsw are in microg/L.

See also

Examples
>>>>>>> refs/remotes/origin/main

  PEC_sw_exposit_runoff(500, Koc = 150)
#> $perc_runoff
#> dissolved     bound 
#>     0.248     0.001 
#> 
#> $runoff
#>           dissolved   bound   total
#> No buffer     1.240 0.00500 1.24500
#> 5 m           0.744 0.00300 0.74700
#> 10 m          0.496 0.00075 0.49675
#> 20 m          0.248 0.00025 0.24825
#> 
#> $PEC_sw_runoff
#>           dissolved       bound    total
#> No buffer  4.769231 0.019230769 4.788462
#> 5 m        4.133333 0.016666667 4.150000
#> 10 m       3.542857 0.005357143 3.548214
#> 20 m       2.480000 0.002500000 2.482500
#> 
  PEC_sw_exposit_runoff(600, Koc = 10000, DT50 = 195, exposit = "3.01a")
#> $perc_runoff
#> dissolved     bound 
#>     0.037     0.159 
#> 
#> $runoff
#>            dissolved      bound      total
#> No buffer 0.21964521 0.94388078 1.16352600
#> 3 m       0.16473391 0.66071655 0.82545046
#> 5 m       0.13178713 0.56632847 0.69811560
#> 6 m       0.12080487 0.42474635 0.54555122
#> 10 m      0.08785809 0.14158212 0.22944020
#> 20 m      0.04392904 0.04719404 0.09112308
#> 
#> $PEC_sw_runoff
#>           dissolved     bound     total
#> No buffer 0.8447893 3.6303107 4.4751000
#> 3 m       0.7844472 3.1462693 3.9307165
#> 5 m       0.7321507 3.1462693 3.8784200
#> 6 m       0.7106169 2.4985080 3.2091248
#> 10 m      0.6275578 1.0113008 1.6388586
#> 20 m      0.4392904 0.4719404 0.9112308
#> 

<<<<<<< HEAD

=======

>>>>>>> refs/remotes/origin/main <<<<<<< HEAD ======= >>>>>>> refs/remotes/origin/main