Calculate PEC surface water due to runoff and erosion as in Exposit 3

This is a reimplementation of the calculation described in the Exposit 3.01 spreadsheet file, in the worksheet "Konzept Runoff". Calculation of sediment PEC values is not implemented.

PEC_sw_exposit_runoff(rate, Koc, DT50 = Inf, t_runoff = 3,
  exposit_reduction_version = c("3.01", "2.0"), V_ditch = 30,
  V_event = 100)

Arguments

rate
The application rate in g/ha

Koc
The sorption coefficient to soil organic carbon

DT50
The soil half-life in days

t_runoff
The time between application and the runoff event, where degradation occurs, in days

exposit_reduction_version
The version of the reduction factors to be used

V_ditch
The volume of the ditch is assumed to be 1 m * 100 m * 30 cm = 30 m3

V_event
The unreduced runoff volume, equivalent to 10 mm precipitation on 1 ha

Source

Excel 3.01 spreadsheet available from https://www.bvl.bund.de/DE/04_Pflanzenschutzmittel/03_Antragsteller/04_Zulassungsverfahren/07_Naturhaushalt/psm_naturhaush_node.html#doc1400590bodyText3

Value

A list containing the following components

perc_runoff

The runoff percentages for dissolved and bound substance

runoff

A matrix containing dissolved and bound input for the different distances

PEC_sw_runoff

A matrix containing PEC values for dissolved and bound substance for the different distances. If the rate was given in g/ha, the PECsw are in microg/L.

See also

Examples

  PEC_sw_exposit_runoff(500, 150)

#> $perc_runoff
#> dissolved     bound 
#>     0.248     0.001 
#> 
#> $runoff
#>           dissolved   bound   total
#> No buffer     1.240 0.00500 1.24500
#> 5 m           0.744 0.00300 0.74700
#> 10 m          0.496 0.00075 0.49675
#> 20 m          0.248 0.00025 0.24825
#> 
#> $PEC_sw_runoff
#>           dissolved       bound    total
#> No buffer  4.769231 0.019230769 4.788462
#> 5 m        4.133333 0.016666667 4.150000
#> 10 m       3.542857 0.005357143 3.548214
#> 20 m       2.480000 0.002500000 2.482500
#>