Calculate PEC surface water at FOCUS Step 1

This is a reimplementation of the FOCUS Step 1 and 2 calculator version 3.2, authored by Michael Klein, in R. Note that results for multiple applications should be compared to the corresponding results for a single application. At current, this is not done automatically in this implementation. Only Step 1 PECs are calculated. However, input files can be generated that are suitable as input for the FOCUS calculator.

Usage

PEC_sw_focus(
  parent,
  rate,
  n = 1,
  i = NA,
  comment = "",
  met = NULL,
  f_drift = NA,
  f_rd = 0.1,
  scenario = FOCUS_Step_12_scenarios$names,
  region = c("n", "s"),
  season = c(NA, "of", "mm", "js"),
  interception = c("no interception", "minimal crop cover", "average crop cover",
    "full canopy"),
  met_form_water = TRUE,
  txt_file = "pesticide.txt",
  overwrite = FALSE,
  append = FALSE
)

Arguments

parent

A list containing substance specific parameters, e.g. conveniently generated by chent_focus_sw.

rate

The application rate in g/ha. Overriden when applications are given explicitly

n

The number of applications

i

The application interval

comment

A comment for the input file

met

A list containing metabolite specific parameters. e.g. conveniently generated by chent_focus_sw. If not NULL, the PEC is calculated for this compound, not the parent.

f_drift

The fraction of the application rate reaching the waterbody via drift. If NA, this is derived from the scenario name and the number of applications via the drift data defined by the FOCUS_Step_12_scenarios

f_rd

The fraction of the amount applied reaching the waterbody via runoff/drainage. At Step 1, it is assumed to be 10%, be it the parent or a metabolite

scenario

The name of the scenario. Must be one of the scenario names given in FOCUS_Step_12_scenarios

region

'n' for Northern Europe or 's' for Southern Europe. If NA, only Step 1 PECsw are calculated

season

'of' for October to February, 'mm' for March to May, and 'js' for June to September. If NA, only step 1 PECsw are calculated

interception

One of 'no interception' (default), 'minimal crop cover', 'average crop cover' or 'full canopy'

met_form_water

Should the metabolite formation in water be taken into account? This can be switched off to check the influence and to compare with previous versions of the Steps 12 calculator

txt_file

the name, and potentially the full path to the Steps.12 input text file to which the specification of the run(s) should be written

overwrite

Should an existing file a the location specified in txt_file be overwritten? Only takes effect if append is FALSE.

append

Should the input text file be appended, if it exists?

Note

The formulas for input to the waterbody via runoff/drainage of the parent and subsequent formation of the metabolite in water is not documented in the model description coming with the calculator. As one would expect, this appears to be (as we get the same results) calculated by multiplying the application rate with the molar weight correction and the formation fraction in water/sediment systems.

Step 2 is not implemented.

References

FOCUS (2014) Generic guidance for Surface Water Scenarios (version 1.4). FOrum for the Co-ordination of pesticde fate models and their USe. http://esdac.jrc.ec.europa.eu/public_path/projects_data/focus/sw/docs/Generic%20FOCUS_SWS_vc1.4.pdf

Website of the Steps 1 and 2 calculator at the Joint Research Center of the European Union: http://esdac.jrc.ec.europa.eu/projects/stepsonetwo

Examples

# Parent only
dummy_1 <- chent_focus_sw("Dummy 1", cwsat = 6000, DT50_ws = 6, Koc = 344.8)
PEC_sw_focus(dummy_1, 3000, f_drift = 0)
#> $f_drift
#> [1] 0
#> 
#> $eq_rate_drift_s
#> [1] 3000
#> 
#> $eq_rate_rd_s
#> [1] 3000
#> 
#> $eq_rate_rd_parent_s
#> [1] NA
#> 
#> $input_drift_s
#> [1] 0
#> 
#> $input_rd_s
#> [1] 300
#> 
#> $f_rd_sw
#> [1] 0.6850566
#> 
#> $f_rd_sed
#> [1] 0.3149434
#> 
#> $PEC
#>      type
#> Time         PECsw   TWAECsw       PECsed  TWAECsed
#>   0   6.850566e+02        NA 2.362075e+03        NA
#>   1   6.103161e+02 647.68635 2.104370e+03 2233.2225
#>   2   5.437298e+02 612.03420 1.874780e+03 2110.2939
#>   4   4.315586e+02 548.76030 1.488014e+03 1892.1255
#>   7   3.051580e+02 469.88375 1.052185e+03 1620.1592
#>   14  1.359325e+02 339.57370 4.686951e+02 1170.8501
#>   21  6.055102e+01 257.45458 2.087799e+02  887.7034
#>   28  2.697241e+01 203.47173 9.300089e+01  701.5705
#>   42  5.352005e+00 140.10377 1.845371e+01  483.0778
#>   50  2.123945e+00 118.24602 7.323361e+00  407.7123
#>   100 6.585062e-03  59.30629 2.270529e-02  204.4881
#> 
#> $PEC_sw_max
#> [1] 685.0566
#> 
#> $PEC_sed_max
#> [1] 2362.075
#> 

# Metabolite
new_dummy <- chent_focus_sw("New Dummy", mw = 250, Koc = 100)
M1 <- chent_focus_sw("M1", mw = 100, cwsat = 100, DT50_ws = 100, Koc = 50,
  max_ws = 0, max_soil = 0.5)
PEC_sw_focus(new_dummy, 1000, scenario = "cereals, winter", met = M1)
#> $f_drift
#> [1] 0.02759
#> 
#> $eq_rate_drift_s
#> [1] 0
#> 
#> $eq_rate_rd_s
#> [1] 200
#> 
#> $eq_rate_rd_parent_s
#> [1] 0
#> 
#> $input_drift_s
#> [1] 0
#> 
#> $input_rd_s
#> [1] 20
#> 
#> $f_rd_sw
#> [1] 0.9375
#> 
#> $f_rd_sed
#> [1] 0.0625
#> 
#> $PEC
#>      type
#> Time     PECsw  TWAECsw   PECsed TWAECsed
#>   0   62.50000       NA 31.25000       NA
#>   1   62.06828 62.28414 31.03414 31.14207
#>   2   61.63954 62.06890 30.81977 31.03445
#>   4   60.79093 61.64158 30.39547 30.82079
#>   7   59.53987 61.00800 29.76994 30.50400
#>   14  56.71995 59.56326 28.35997 29.78163
#>   21  54.03358 58.16414 27.01679 29.08207
#>   28  51.47444 56.80902 25.73722 28.40451
#>   42  46.71404 54.22460 23.35702 27.11230
#>   50  44.19417 52.81945 22.09709 26.40973
#>   100 31.25000 45.08422 15.62500 22.54211
#> 
#> $PEC_sw_max
#> [1] 62.5
#> 
#> $PEC_sed_max
#> [1] 31.25
#>