Set up runs for FOCUS PELMO

Per default, the runs are not only set up but also executed with FOCUS PELMO, the results are processed and returned. Currently, only FOCUS PELMO as installed on Linux (or other Unix systems) using the install_PELMO from the PELMO.installeR package maintained on github is supported. In such installations, FOCUS PELMO is installed into the package installation directory of PELMO.installeR and run using wine.

PELMO_runs(runs, psm_dir = ".", version = "5.5.3", PELMO_base = "auto",
  execute = TRUE, cores = getOption("mc.cores", 2L), evaluate = TRUE,
  overwrite = FALSE)

run_PELMO(runs, version = "5.5.3", PELMO_base = "auto",
  cores = getOption("mc.cores", 2L))

evaluate_PELMO(runs, version = "5.5.3", PELMO_base = "auto")

Arguments

runs

A list of lists. Each inner lists has an element named 'psm' that holds the psm string, and elements named using three letter crop acronyms, as used in FOCUS_PELMO_crops, that hold character vectors of three letter scenario acronyms as used in FOCUS_GW_scenarios_2012.

psm_dir

The directory where the psm files are located

version

The FOCUS PELMO version

PELMO_base

Where the FOCUS PELMO installation is located

execute

Should PELMO be executed directly?

cores

The number of cores to execute PELMO runs in parallel

evaluate

Should the results be returned?

overwrite

Should an existing run directories be overwritten

References

PELMO.installeR https://jranke.github.io/PELMO.installeR

Wine https://winehq.org