Deposition from spray drift expressed as percent of the applied dose as published by the JKI

Deposition from spray drift expressed as percent of the applied dose as published by the German Julius-Kühn Institute (JKI).

Format

A list currently containing matrices with spray drift percentage data for field crops (Ackerbau), and Pome/stone fruit, early and late (Obstbau frueh, spaet).

Source

JKI (2010) Spreadsheet 'Tabelle der Abdrifteckwerte.xls', retrieved from http://www.jki.bund.de/no_cache/de/startseite/institute/anwendungstechnik/abdrift-eckwerte.html on 2015-06-11

Details

The data were extracted from the spreadsheet cited below using the R code given in the example section. The spreadsheet is not included in the package as its licence is not clear.

Examples

# NOT RUN {
  # This is the code that was used to extract the data
  library(readxl)
  abdrift_path <- "inst/extdata/Tabelle der Abdrifteckwerte.xls"
  JKI_crops <- c("Ackerbau", "Obstbau frueh", "Obstbau spaet")
  names(JKI_crops) <- c("Field crops", "Pome/stone fruit, early", "Pome/stone fruit, late")
  drift_data_JKI <- list()

  for (n in 1:8) {
    drift_data_raw <- read_excel(abdrift_path, sheet = n + 1, skip = 2)
    drift_data <- as.matrix(drift_data_raw[1:9, 2:4])
    dimnames(drift_data) <- list(distance = as.integer(drift_data_raw[1:9, 1]),
                                            crop = JKI_crops)
    drift_data_JKI[[n]] <- drift_data
  }
  save(drift_data_JKI, file = "data/drift_data_JKI.RData")
# }

# And this is the resulting data
drift_data_JKI

#> [[1]]
#>         crop
#> distance Ackerbau Obstbau frueh Obstbau spaet Weinbau Hopfenbau
#>       1      2.77            NA            NA      NA        NA
#>       3      0.95         29.20         15.73    8.02     19.33
#>       5      0.57         19.89          8.41    3.62     11.57
#>       10     0.29         11.81          3.60    1.23      5.77
#>       15     0.20          5.55          1.81    0.65      3.84
#>       20     0.15          2.77          1.09    0.42      1.79
#>       30     0.10          1.04          0.54    0.22      0.56
#>       40     0.07          0.52          0.32    0.14      0.25
#>       50     0.06          0.30          0.22    0.10      0.13
#>         crop
#> distance Flaechenkulturen > 900 l/ha Tragbar < 50 cm
#>       1                        4.440            2.77
#>       3                           NA              NA
#>       5                        0.180            0.57
#>       10                       0.050            0.29
#>       15                       0.020            0.20
#>       20                       0.012            0.15
#>       30                       0.005            0.10
#>       40                       0.003            0.07
#>       50                       0.002            0.06
#>         crop
#> distance Tragbar < 50 cm mit Spritzschirm Tragbar > 50 cm Gleisanlagen
#>       1                             0.039              NA           NA
#>       3                                NA            8.02  0.018721696
#>       5                             0.005            3.62  0.014363896
#>       10                               NA            1.23  0.010026007
#>       15                               NA            0.65  0.008124366
#>       20                               NA            0.42  0.006998158
#>       30                               NA            0.22  0.005670811
#>       40                               NA            0.14           NA
#>       50                               NA            0.10  0.004350831
#> 
#> [[2]]
#>         crop
#> distance Ackerbau Obstbau frueh Obstbau spaet Weinbau Hopfenbau
#>       1      2.38            NA            NA      NA        NA
#>       3      0.79         25.53         12.13    7.23     17.73
#>       5      0.47         16.87          6.81    3.22      9.60
#>       10     0.24          9.61          3.11    1.07      4.18
#>       15     0.16          5.61          1.58    0.56      2.57
#>       20     0.12          2.59          0.90    0.36      1.21
#>       30     0.08          0.87          0.40    0.19      0.38
#>       40     0.06          0.40          0.23    0.12      0.17
#>       50     0.05          0.22          0.15    0.08      0.09
#>         crop
#> distance Flaechenkulturen > 900 l/ha Tragbar < 50 cm
#>       1                        3.780            2.38
#>       3                           NA              NA
#>       5                        0.160            0.47
#>       10                       0.040            0.24
#>       15                       0.020            0.16
#>       20                       0.011            0.12
#>       30                       0.005            0.08
#>       40                       0.003            0.06
#>       50                       0.002            0.05
#>         crop
#> distance Tragbar < 50 cm mit Spritzschirm Tragbar > 50 cm Gleisanlagen
#>       1                                NA          0.1900           NA
#>       3                              7.23              NA        2.120
#>       5                              3.22          0.0100        0.470
#>       10                             1.07          0.0030        0.060
#>       15                             0.56          0.0010        0.020
#>       20                             0.36          0.0008        0.008
#>       30                             0.19              NA           NA
#>       40                             0.12              NA           NA
#>       50                             0.08              NA           NA
#> 
#> [[3]]
#>         crop
#> distance Ackerbau Obstbau frueh Obstbau spaet Weinbau Hopfenbau
#>       1      2.01            NA            NA      NA        NA
#>       3      0.68         23.96         11.01    6.90     15.93
#>       5      0.41         15.79          6.04    3.07      8.57
#>       10     0.20          8.96          2.67    1.02      3.70
#>       15     0.14          4.24          1.39    0.54      2.26
#>       20     0.10          2.01          0.80    0.34      1.05
#>       30     0.07          0.70          0.36    0.18      0.34
#>       40     0.05          0.33          0.21    0.11      0.15
#>       50     0.04          0.19          0.13    0.08      0.08
#>         crop
#> distance Flaechenkulturen > 900 l/ha Tragbar < 50 cm
#>       1                        3.420            2.01
#>       3                           NA              NA
#>       5                        0.150            0.41
#>       10                       0.040            0.20
#>       15                       0.020            0.14
#>       20                       0.010            0.10
#>       30                       0.004            0.07
#>       40                       0.003            0.05
#>       50                       0.002            0.04
#>         crop
#> distance Tragbar < 50 cm mit Spritzschirm Tragbar > 50 cm Gleisanlagen
#>       1                                NA          0.1400           NA
#>       3                              6.90              NA        1.590
#>       5                              3.07          0.0100        0.360
#>       10                             1.02          0.0020        0.050
#>       15                             0.54          0.0010        0.010
#>       20                             0.34          0.0006        0.006
#>       30                             0.18              NA           NA
#>       40                             0.11              NA           NA
#>       50                             0.08              NA           NA
#> 
#> [[4]]
#>         crop
#> distance Ackerbau Obstbau frueh Obstbau spaet Weinbau Hopfenbau
#>       1      1.85            NA            NA      NA        NA
#>       3      0.62         23.61         10.12    6.71     15.38
#>       5      0.38         15.42          5.60    2.99      8.26
#>       10     0.19          8.66          2.50    0.99      3.55
#>       15     0.13          4.01          1.28    0.52      2.17
#>       20     0.10          1.89          0.75    0.33      0.93
#>       30     0.06          0.66          0.35    0.17      0.31
#>       40     0.05          0.31          0.20    0.11      0.14
#>       50     0.04          0.17          0.13    0.08      0.08
#>         crop
#> distance Flaechenkulturen > 900 l/ha Tragbar < 50 cm
#>       1                        2.290            1.85
#>       3                           NA              NA
#>       5                        0.120            0.38
#>       10                       0.030            0.19
#>       15                       0.020            0.13
#>       20                       0.009            0.10
#>       30                       0.004            0.06
#>       40                       0.002            0.05
#>       50                       0.002            0.04
#>         crop
#> distance Tragbar < 50 cm mit Spritzschirm Tragbar > 50 cm Gleisanlagen
#>       1                                NA          0.1300           NA
#>       3                              6.71              NA        1.360
#>       5                              2.99          0.0100        0.310
#>       10                             0.99          0.0020        0.040
#>       15                             0.52          0.0010        0.010
#>       20                             0.33          0.0005        0.005
#>       30                             0.17              NA           NA
#>       40                             0.11              NA           NA
#>       50                             0.08              NA           NA
#> 
#> [[5]]
#>         crop
#> distance Ackerbau Obstbau frueh Obstbau spaet Weinbau Hopfenbau
#>       1      1.75            NA            NA      NA        NA
#>       3      0.59         23.12          9.74    6.59     15.12
#>       5      0.36         15.06          5.41    2.93      7.99
#>       10     0.18          8.42          2.43    0.98      3.36
#>       15     0.12          3.83          1.24    0.51      2.03
#>       20     0.09          1.81          0.72    0.33      0.88
#>       30     0.06          0.63          0.34    0.17      0.29
#>       40     0.05          0.30          0.20    0.11      0.14
#>       50     0.04          0.17          0.13    0.08      0.07
#>         crop
#> distance Flaechenkulturen > 900 l/ha Tragbar < 50 cm
#>       1                        2.120            1.75
#>       3                           NA              NA
#>       5                        0.110            0.36
#>       10                       0.030            0.18
#>       15                       0.010            0.12
#>       20                       0.008            0.09
#>       30                       0.004            0.06
#>       40                       0.002            0.05
#>       50                       0.002            0.04
#>         crop
#> distance Tragbar < 50 cm mit Spritzschirm Tragbar > 50 cm Gleisanlagen
#>       1                                NA          0.1200           NA
#>       3                              6.59              NA        1.190
#>       5                              2.93          0.0100        0.270
#>       10                             0.98          0.0020        0.040
#>       15                             0.51          0.0010        0.010
#>       20                             0.33          0.0005        0.005
#>       30                             0.17              NA           NA
#>       40                             0.11              NA           NA
#>       50                             0.08              NA           NA
#> 
#> [[6]]
#>         crop
#> distance Ackerbau Obstbau frueh Obstbau spaet Weinbau Hopfenbau
#>       1      1.64            NA            NA      NA        NA
#>       3      0.56         22.76          9.21    6.41     14.90
#>       5      0.34         14.64          5.18    2.85      7.79
#>       10     0.17          8.04          2.38    0.95      3.23
#>       15     0.11          3.71          1.20    0.50      1.93
#>       20     0.09          1.75          0.68    0.32      0.83
#>       30     0.06          0.61          0.31    0.17      0.28
#>       40     0.04          0.29          0.17    0.11      0.13
#>       50     0.03          0.16          0.11    0.07      0.07
#>         crop
#> distance Flaechenkulturen > 900 l/ha Tragbar < 50 cm
#>       1                        1.980            1.64
#>       3                           NA              NA
#>       5                        0.100            0.34
#>       10                       0.030            0.17
#>       15                       0.010            0.11
#>       20                       0.008            0.09
#>       30                       0.004            0.06
#>       40                       0.002            0.04
#>       50                       0.001            0.03
#>         crop
#> distance Tragbar < 50 cm mit Spritzschirm Tragbar > 50 cm Gleisanlagen
#>       1                                NA          0.1100           NA
#>       3                              6.41              NA        1.070
#>       5                              2.85          0.0100        0.250
#>       10                             0.95          0.0020        0.030
#>       15                             0.50          0.0010        0.010
#>       20                             0.32          0.0004        0.005
#>       30                             0.17              NA           NA
#>       40                             0.11              NA           NA
#>       50                             0.07              NA           NA
#> 
#> [[7]]
#>         crop
#> distance Ackerbau Obstbau frueh Obstbau spaet Weinbau Hopfenbau
#>       1      1.61            NA            NA      NA        NA
#>       3      0.55         22.69          9.10    6.33     14.63
#>       5      0.33         14.45          5.11    2.81      7.60
#>       10     0.17          7.83          2.33    0.94      3.13
#>       15     0.11          3.62          1.20    0.49      1.86
#>       20     0.08          1.71          0.67    0.31      0.81
#>       30     0.06          0.60          0.30    0.16      0.26
#>       40     0.04          0.28          0.17    0.10      0.12
#>       50     0.03          0.16          0.11    0.07      0.06
#>         crop
#> distance Flaechenkulturen > 900 l/ha Tragbar < 50 cm
#>       1                        1.930            1.61
#>       3                           NA              NA
#>       5                        0.100            0.33
#>       10                       0.030            0.17
#>       15                       0.010            0.11
#>       20                       0.008            0.08
#>       30                       0.004            0.06
#>       40                       0.002            0.04
#>       50                       0.001            0.03
#>         crop
#> distance Tragbar < 50 cm mit Spritzschirm Tragbar > 50 cm Gleisanlagen
#>       1                                NA          0.1100           NA
#>       3                              6.33              NA        0.960
#>       5                              2.81          0.0100        0.230
#>       10                             0.94          0.0020        0.030
#>       15                             0.49          0.0010        0.010
#>       20                             0.31          0.0004        0.004
#>       30                             0.16              NA           NA
#>       40                             0.10              NA           NA
#>       50                             0.07              NA           NA
#> 
#> [[8]]
#>         crop
#> distance Ackerbau Obstbau frueh Obstbau spaet Weinbau Hopfenbau
#>       1      1.52            NA            NA      NA        NA
#>       3      0.52         22.24          8.66    6.26     13.53
#>       5      0.31         14.09          4.92    2.78      7.15
#>       10     0.16          7.58          2.29    0.93      3.01
#>       15     0.11          3.48          1.14    0.49      1.82
#>       20     0.08          1.65          0.65    0.31      0.78
#>       30     0.05          0.57          0.29    0.16      0.25
#>       40     0.04          0.27          0.16    0.10      0.12
#>       50     0.03          0.15          0.11    0.07      0.06
#>         crop
#> distance Flaechenkulturen > 900 l/ha Tragbar < 50 cm
#>       1                        1.640            1.52
#>       3                           NA              NA
#>       5                        0.090            0.31
#>       10                       0.020            0.16
#>       15                       0.010            0.11
#>       20                       0.007            0.08
#>       30                       0.003            0.05
#>       40                       0.002            0.04
#>       50                       0.001            0.03
#>         crop
#> distance Tragbar < 50 cm mit Spritzschirm Tragbar > 50 cm Gleisanlagen
#>       1                                NA           1e-01           NA
#>       3                              6.26              NA        0.890
#>       5                              2.78           1e-02        0.210
#>       10                             0.93           1e-03        0.030
#>       15                             0.49           1e-03        0.010
#>       20                             0.31           4e-04        0.004
#>       30                             0.16              NA           NA
#>       40                             0.10              NA           NA
#>       50                             0.07              NA           NA
#>