Read TOXSWA surface water concentrations

Read TOXSWA hourly concentrations of a chemical substance in a specific segment of a TOXSWA surface water body. Per default, the data for the last segment are imported. As TOXSWA 4 reports the values at the end of the hour (ConLiqWatLayCur) in its summary file, we use this value as well instead of the hourly averages (ConLiqWatLay). In TOXSWA 5.5.3 this variable was renamed to ConLiqWatLay in the out file.

read.TOXSWA_cwa(filename, basedir = ".", zipfile = NULL,
  segment = "last", substance = "parent", total = FALSE,
  windows = NULL, thresholds = NULL)

Arguments

filename
The filename of the cwa file (TOXSWA 2.x.y or similar) or the out file when using FOCUS TOXSWA 4 (i.e. TOXSWA 4.4.2) or higher.

basedir
The path to the directory where the cwa file resides.

zipfile
Optional path to a zip file containing the cwa file.

segment
The segment for which the data should be read. Either "last", or the segment number.

substance
For .out files, the default value "parent" leads to reading concentrations of the parent compound. Alternatively, the substance of interested can be selected by its code name.

total
Set this to TRUE in order to read total concentrations as well. This is only necessary for .out files as generated by TOXSWA 4.4.2 or similar, not for .cwa files. For .cwa files, the total concentration is always read as well.

windows
Numeric vector of width of moving windows in days, for calculating maximum time weighted average concentrations and areas under the curve.

thresholds
Numeric vector of threshold concentrations in µg/L for generating event statistics.

Value

An instance of an R6 object of class TOXSWA_cwa.

Examples

H_sw_D4_pond  <- read.TOXSWA_cwa("00001p_pa.cwa",
                                 basedir = "SwashProjects/project_H_sw/TOXSWA",
                                 zipfile = system.file("testdata/SwashProjects.zip",
                                                       package = "pfm"))